% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance_transformation.R
\name{abundance_transformation}
\alias{abundance_transformation}
\title{Transform a numeric master_daten frame}
\usage{
abundance_transformation(
  master.table,
  transformation = c("percentage", "proportion", "scale"),
  internal_standard_peak = NULL,
  internal_standard_amount = NULL
)
}
\arguments{
\item{master.table}{A master table, for which the abundance values should be
transformed.}

\item{transformation}{The type of transformation to perform.
Can be one of "percentage", "proportion", or "scale".
Defaults to "percentage".}

\item{internal_standard_peak}{A character specifying the Peak (e.g. "P2") that
contians the internal standard added to the samples, to be used
\code{transformation} is set to "scale".}

\item{internal_standard_amount}{A numeric value, indicating the known amount
of internal standard that was added to the samples.}
}
\description{
This function takes a master table and transforms the peak abundance values
within each sample, based on the specified transformation option.
}
