% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comps_info.R
\name{add_comps_info}
\alias{add_comps_info}
\title{Add the compounds' information to the aligned data sets within a list}
\usage{
add_comps_info(samples.list, comps.info.list)
}
\arguments{
\item{samples.list}{List of aligned data sets as obtained with
\link{adjust_abundance}}

\item{comps.info.list}{List of data frames with the compounds' information,
as obtained with \link{get_chc_info}.}
}
\description{
Add the compounds' information to the aligned data sets within a list
}
