% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_abundance.R
\name{adjust_abundance}
\alias{adjust_abundance}
\title{Adjust abundance values within samples and display diagnostic plots}
\usage{
adjust_abundance(aligned_data, std.info, plot = T)
}
\arguments{
\item{aligned_data}{Aligned data set, as obtained with \link{recalculate_meanRT}.}

\item{std.info}{Data frame with the standards information, as obtained
from \link{shape_hcstd_info}.
It must contains columns "mean_RT" and "area_correction".}

\item{plot}{Logical value indicating whether to generate plots illustrating
the change in the abundance data of the samples (default: TRUE).
This generates a bar plot per sample showing the abundance on y-axis and
mean_RT on x-axis.
The plots are divided in two facets, to show the un-adjusted and adjusted
versions of the peaks' abundance within a sample.}
}
\description{
This function can be used to correct for the difference in performance of
the total count of ions for the different peaks, in the case of the analysis
of hydrocarbon mixtures. The adjustment is based on the total ion counts for
the peaks detected in n-alkane analytical standard solutions.
}
\examples{

# Adjusting the abundance for a single data set
IW_data <- corrected_samples_list$`Winter_In-hive workers_A. m. mellifera`
  adjusted_IW <- adjust_abundance(aligned_data = IW_data
                                        , std.info = std_info)

# Adjusting the abundance for several data sets within a list
adjusted_samples_list <- corrected_samples_list |>
  lapply(adjust_abundance, std.info = std_info)

}
