% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aligned_samples_data_list}
\alias{aligned_samples_data_list}
\title{GC/MS data of honey bees cuticular hydrocarbons aligned by group
(In- and out-hive workers).}
\format{
\subsection{}{

A list with two GCalignR objects:
\describe{
\item{Winter_In-hive workers_A. m. mellifera}{Alignment of the GC/MS
integration dara of the in-hive workers.}
\item{Winter_Out-hive workers_A. m. mellifera}{Alignment of the GC/MS
integration dara of the out-hive workers.}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
aligned_samples_data_list
}
\description{
GC/MS data of honey bees cuticular hydrocarbons aligned by group
(In- and out-hive workers).
}
\keyword{datasets}
