% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aligned_standards}
\alias{aligned_standards}
\title{GC/MS data of n-alkane analytical standard solutions aligned}
\format{
\subsection{}{

A GCalignR object.
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
aligned_standards
}
\description{
GC/MS data of n-alkane analytical standard solutions aligned
}
\keyword{datasets}
