% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_df.R
\name{area_df}
\alias{area_df}
\title{Extract aligned area data frames}
\usage{
area_df(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by align_chromatogram2.}
}
\description{
A function to obtain the area aligned data frame from an
aligned data set, as produced by align_chromatogram2.
It automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
\examples{

# Extract area data frame from a single aligned data set
Winter_IW <-
  aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`

area_winter_IW <- area_df(Winter_IW)

# Extract area data frame from several aligned data sets within a list
samples_list_area <- aligned_samples_data_list |>
  lapply(area_df)

}
