% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_norm.R
\name{area_norm}
\alias{area_norm}
\title{Normalize aligned area data frame}
\usage{
area_norm(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by \link{align_chromatograms2}}
}
\description{
Function to normalize the aligned area data frame produced by
\link{align_chromatograms2}.

It is a wrapper around \link[GCalignR:norm_peaks]{GCalignR::norm_peaks}.

It facilitates the usage of the function with \link{lapply}, but its main purpose
is to be called within \link{diagnostic_heatmap}.
By default set the rt_col_name to "RT", and conc_col_name to "Area".

Automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
\examples{

# Normalize a single aligned data set
Winter_IW <-
  aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`

norm_winter_IW <- area_norm(Winter_IW)

# Normalize several aligned data sets within a list
samples_area_norm_list <- aligned_samples_data_list |>
  lapply(area_norm)

}
