% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_duplicated_compounds.R
\name{assess_duplicated_compounds}
\alias{assess_duplicated_compounds}
\title{Assess the presence of duplicated compounds among groups}
\usage{
assess_duplicated_compounds(group.tables.list, plot = T)
}
\arguments{
\item{group.tables.list}{List of group tables extracted from the master table
with \link{retrieve_group_tables}.}

\item{plot}{Logical value indicating whether to generate density distribution
plots for each compound (default: TRUE)}
}
\description{
This function takes in a list of group tables extracted from a
master table and assesses the presence of duplicated compounds among groups.
}
