% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_master_table.R
\name{build_master_table}
\alias{build_master_table}
\title{Build a master table from several group tables}
\usage{
build_master_table(tables.list)
}
\arguments{
\item{tables.list}{List of group tables, as obtained from
\link{shape_aligned_table}.
The group tables must contain an integer RI column, indicating the retention
index of the peaks (rows).}
}
\description{
Function to build a master table from a list of group tables.
It relies on the base::merge function to fuse the group tables together
into the master table.
}
