% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comps_id_std}
\alias{comps_id_std}
\title{Compound identification of the n-alkanes in the analytical standard
solutions (04070-1ML and 04071-5ML, Sigma-Aldrich).}
\format{
\subsection{}{

A tibble data frame with 59 rows and 5 columns:
\describe{
\item{Peak}{Peak labels}
\item{Compound}{ID of the compound(s) contained by a peak}
\item{mean_RT}{Mean RT of the peak among samples}
\item{H2005}{Sample corresponding to an n-alkane 04071-5ML analytical
standard solution.}
\item{L2005}{Sample corresponding to an n-alkane 04070-1ML analytical
standard solution.}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
comps_id_std
}
\description{
Compound identification of the n-alkanes in the analytical standard
solutions (04070-1ML and 04071-5ML, Sigma-Aldrich).
}
\keyword{datasets}
