% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comps_info_list}
\alias{comps_info_list}
\title{Information of hydrocarbons in the cuticular hydrocarbon extracts of the
honey bees.}
\format{
\subsection{}{

A list with two tibble data frames:
\describe{
\item{Winter_In-hive workers_A. m. mellifera}{
Information of compounds in the cuticular hydrocarbon extracts of the
in-hive workers}
\item{Winter_Out-hive workers_A. m. mellifera}{
Information of compounds in the cuticular hydrocarbon extracts of the
out-hive workers}
Both tibble data frames they contain the labels of each peak (Peak),
the ID of the hydrocarbons in the corresponding peak (Compound), the
hydrocarbon chain length (Chain.length), the hydrocarbon class (Class),
and the position of unsaturations or methyl-branches in the hydrocarbon
chain (Mod.position).
These tibble data frames were generated by applying \code{get_hc_info} on the
tibble data frames contained in \code{comps_id_list}.

}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
comps_info_list
}
\description{
Information of hydrocarbons in the cuticular hydrocarbon extracts of the
honey bees.
}
\keyword{datasets}
