% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corrected_samples_list2}
\alias{corrected_samples_list2}
\title{Corrected alignment of the honey bees cuticular hydrocarbons GC/MS data, with
corrected mean RT.}
\format{
\subsection{}{

A list with two sublists:
\describe{
\item{Winter_In-hive workers_A. m. mellifera}{
List with the RT and Area tibble data frames for the corrected alignment
with recalculated mean RT of the in-hive workers.
}
\item{Winter_Out-hive workers_A. m. mellifera}{
List with the RT and Area tibble data frames for the corrected alignment
with recalculated mean RT of the out-hive workers.
}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
corrected_samples_list2
}
\description{
Corrected alignment of the honey bees cuticular hydrocarbons GC/MS data, with
corrected mean RT.
}
\keyword{datasets}
