% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_heatmap.R
\name{diagnostic_heatmap}
\alias{diagnostic_heatmap}
\title{Plot to evaluate peak alignment}
\usage{
diagnostic_heatmap(data, title)
}
\arguments{
\item{data}{The function can receive different data types from which to produce the
diagnostic plot:

Aligned data, as obtained from \link{align_chromatograms2}, for which you want to
evaluate its alignment. In this case, it calls area_norm to normalize
the aligned area data frame before producing the plot.

An aligned data.frame of normalized abundance (area), as obtained from
area_norm.

A corrected aligned data.frame as obtained from correct_alignment.}

\item{title}{Character string with the text for the title of the plot}
}
\description{
Heat map + dendrogram of the chemical composition of samples within an
aligned data frame. The colors filling the heat map cells represent the
abundance of the compounds (columns), and the dendrogram indicates the
similarity between samples (rows).

The plot facilitates the discovery of miss-alignments.

It uses \link[gplots:heatmap.2]{gplots::heatmap.2} to produce the plot.

If the received data frame contains only on sample, it returns a heatmap,
using ggplot2

The data is transformed to the log(1 + x) before plotting, so the difference
between peaks with low abundance values (i.e x >= 5\%) is observable.
}
\examples{

# Get the diagnostic heat map of a single aligned data set
Winter_IW <-
  aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`

diagnostic_heatmap(data =  Winter_IW
                   , title = "Alignment of IW CHCs"
                   , alignment.type = "automatic")

# Get the diagnostic heat map of several aligned data sets within a list

for (df in names(aligned_samples_data_list)) {
  diagnostic_heatmap(aligned_samples_data_list[[df]]
                     , title = paste0("Alignment of ", df)
                     , alignment.type = "automatic")
}


}
