% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_compounds.R
\name{drop_na_compounds}
\alias{drop_na_compounds}
\title{Remove the peaks that do not contain compounds of interest}
\usage{
drop_na_compounds(filtered_data)
}
\arguments{
\item{filtered_data}{}
}
\description{
Remove the peaks that do not contain compounds of interest
}
