% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_samples_list}
\alias{filtered_samples_list}
\title{GC/MS data of honeybees cuticular hydrocarbons aligned by group}
\format{
\subsection{}{

A list with two GCalignR objects:
\describe{
\item{}{}
\item{}{}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
filtered_samples_list
}
\description{
GC/MS data of honeybees cuticular hydrocarbons aligned by group
}
\keyword{datasets}
