% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_all_peaks.R
\name{fuse_all_peaks}
\alias{fuse_all_peaks}
\title{Fuse peaks to correct miss-alignemnets in a master table}
\usage{
fuse_all_peaks(master.table, fusion.list)
}
\arguments{
\item{master.table}{A master table data frame, as obtained from
\link{build_master_table}.}

\item{fusion.list}{List of data frames specifying the fusions to be performed.}
}
\description{
The function fuses peaks within a master table as specified by
a list. It uses a \link{fuse_peaks} to perform each fusion.
}
