% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_all_peaks2.R
\name{fuse_all_peaks2}
\alias{fuse_all_peaks2}
\title{Fuse peaks}
\usage{
fuse_all_peaks2(df, fusion.list, df_type)
}
\arguments{
\item{df}{A master table data frame, as obtained from
\link{build_master_table}, OR an aligned data frame as obtained from
\link{add_comps_info}}

\item{fusion.list}{List of data frames specifying the fusions to be performed.}

\item{df_type}{type of data frame given as df; either "master.table" OR
"group.table".}
}
\description{
The function fuses peaks within a master table as specified by
a list.
}
