% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuse_peaks.R
\name{fuse_peaks}
\alias{fuse_peaks}
\title{Fuse several peaks together}
\usage{
fuse_peaks(master.table, peaks_to_fuse)
}
\arguments{
\item{master.table}{A master table data frame, as obtained from
\link{build_master_table}.}

\item{peaks_to_fuse}{A character vector specifying the peaks to be fused.}
}
\description{
Function to fuse the indicated peaks into one. It is called by
\link{fuse_all_peaks} to perform each of the peak fusion procedures on a master
table.
}
