% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chc_info.R
\name{get_chc_info}
\alias{get_chc_info}
\title{get the CHC information data frame}
\usage{
get_chc_info(comps.id)
}
\arguments{
\item{comps.id}{}
}
\description{
Function to generate the data frame holding the compound
information for the hydrocarbons present in each sample, within a table
}
