% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daten.R
\name{get_daten}
\alias{get_daten}
\title{Extract the abundance data from  a finished master table}
\usage{
get_daten(master.table)
}
\arguments{
\item{master.table}{A finished master table data frame}
}
\description{
A function to extract the data frame with the peak
abundance per sample from the final master table
}
