% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hc_info.R
\name{get_hc_info}
\alias{get_hc_info}
\title{get information f hydrocarbons}
\usage{
get_hc_info(comps.id)
}
\arguments{
\item{comps.id}{data frame with the compound information for the hydrocarbons found in each
sample.

The data frame should have a \strong{"Compound"} and a \strong{"mean_RT"} columns, plus
a column for every sample.

The Compound column should have the name of the hydrocarbons with the
following format:
C\emph{numeric chain length} _ \emph{class abbreviation} _
\emph{numeric position of unsaturations / methyl-branches}.

As an example the entry for the compound 15-Me C20 would be C20_Me_15-.

The function automatically extracts the information of the compound from the
entry in the \strong{"Compound"} column, and replaces it by the correct name.

If a peak contains more than one compound with unsaturations /
methyl-branches, they must be separated by a ";" (e.g. C20_Me_13-;15-
for a peak containing both 13-Me C20 and 15-Me C20).
In the case of n-alkanes, or a hydrocabron for which the position of its
unsaturation(s) is not known, this must be specified with "NA"
(e.g. C20_ane_NA, or C20_ene_NA)
\subsection{Class abbreviations accepted by the function:}{

n-alkanes: ane
alkene: ene
alkadiene: diene
monomethyl-branched alkanes: Me
Dimethyl-branched alkanes: Dime
Trimethyl-branched alkanes: Trime

The rest of the data frame should be just as it was extracted from the RT
entry of the aligned standards.
}}
}
\description{
Function to generate the data frame holding the compound
information for the hydrocarbons present in each sample, within a table
}
