% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_frequency_filter.R
\name{group_frequency_filter}
\alias{group_frequency_filter}
\title{Remove rare compounds from a group table}
\usage{
group_frequency_filter(group_table, threshold = 0.5)
}
\arguments{
\item{group_table}{A group table data frame, as obtained with
\link{retrieve_group_tables}.}

\item{threshold}{A numeric value, between 0 and 1 (default = 0.5), indicating
the minimum proportion of samples in which a compound should be within the
group table to not be erased from it.}
}
\description{
The function removes from a group table, the compounds that are
rare to that group, regarding to a minimum frequency threshold.
Any compound that has a frequency within the group table below the given
threshold is erased from the data frame of the group table.
}
