% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grouping_info}
\alias{grouping_info}
\title{Group membership information of honeybee samples in the \code{samples_data_list}
data object.}
\format{
\subsection{}{

A tibble data frame with 20 rows and 4 columns:
\describe{
\item{Season}{Season during which the samples were collected (all in
Winter).}
\item{Task}{The type of worker (in- or out-hive) the bees were when
collected.}
\item{Subspecies}{Subspecies of the bees}
\item{Individual}{Individual ID number}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
grouping_info
}
\description{
Group membership information of honeybee samples in the \code{samples_data_list}
data object.
}
\keyword{datasets}
