% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gcms_data.R
\name{import_gcms_data}
\alias{import_gcms_data}
\title{Import GC integration results from CSV  files}
\usage{
import_gcms_data(data.path.list, patterns_2_delete)
}
\arguments{
\item{data.path.list}{list containing the paths of the CSV files from which
to import the integration results of each sample.}

\item{patterns_2_delete}{a character type vector, listing the text  patterns
to be removed from the CSV files' paths, to extract the names of the samples
from them.}
}
\value{
A list of tibble data frames with the integration results
(RT: retention time of each peak, and Area: area under the curve of each
peak).
Each tibble on the list corresponds to the CSV file of an individual
sample.
}
\description{
A function to import the integration results data frames,
as obtained from MassHunter.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
