% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kovats_retention_index.R
\name{kovats_retention_index}
\alias{kovats_retention_index}
\title{Calculate the Kováts retention index}
\usage{
kovats_retention_index(filtered_data, std.info)
}
\arguments{
\item{filtered_data}{An aligned data set list containing data frames with
information about compounds and their peak areas and retention times.}

\item{std.info}{A data frame containing information about the standards
that will be used in the calculation of retention indices, as obtained with
\link{shape_hcstd_info}.}
}
\description{
This function calculates the retention indices for the peaks in a data set
using the Kováts method. The retention indices are calculated based on the
mean retention times of n-alkanes within the samples or in a standard mixture
in case of an n-alkane that is not present in the samples.
This function is used within \code{shape_aligned_table}.
}
