% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_list.R
\name{mg_list}
\alias{mg_list}
\title{group samples within a list}
\usage{
mg_list(sample.info, group.label, samples.data.list)
}
\arguments{
\item{sample.info}{A data frame indicating the groups to which the samples belong}

\item{group.label}{A character indicating the name of the column of sample.info that contains
the group names by which the samples should be grouped}

\item{samples.data.list}{List of data frames with integration results per sample, as produced by
\link{import_mh_data}}
}
\description{
A function to Group integration data frames of samples, based
on unique group labels, within the list containing the data frames.
}
\examples{

library(tidyr)

grouping_info <- grouping_info |>
  unite(group_label
        , where(is.factor)
        , sep = "_"
        , remove = FALSE)

# Nest the data frames in sublists by group
samples_data_list <- mg_list(sample.info = grouping_info
                             , group.label = "group_label"
                             , samples.data.list = samples_data_list)

}
