% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_group_tables.R
\name{retrieve_group_tables}
\alias{retrieve_group_tables}
\title{Retrieve the group tables from the master table}
\usage{
retrieve_group_tables(group.label, master.table, grouping.info)
}
\arguments{
\item{group.label}{Character indicating the column of grouping.info
that contains the labels of the groups that define the group tables.}

\item{master.table}{master table data frames, as obtained with
\link{build_master_table}.}

\item{grouping.info}{Data frame with information of the samples, within which
the group labels can be found.}
}
\description{
Retrieve the group tables from the master table
}
