% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_group_tables2.R
\name{retrieve_group_tables2}
\alias{retrieve_group_tables2}
\title{Retrieve the group tables from the master table}
\usage{
retrieve_group_tables2(group.label, master.table, grouping.info)
}
\arguments{
\item{group.label}{Character indicating the column of grouping.info
that contains the labels of the groups that define the group tables.
It is meant to be use din a pipe line with \link{assess_duplicated_compounds} or
\link{group_frequency_filter}.}

\item{master.table}{master table data frames, as obtained with
\link{build_master_table}.}

\item{grouping.info}{Data frame with information of the samples, within which
the group labels can be found.}
}
\description{
Retrieve the group tables from the master table
}
