% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_df.R
\name{RT_df}
\alias{RT_df}
\title{Extract aligned retention time data frames}
\usage{
RT_df(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by align_chromatogram2}
}
\description{
A function to obtain the retention time (RT) aligned data frame
from an aligned data set, as produced by align_chromatogram2.
It automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
