% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_data_list}
\alias{samples_data_list}
\title{GC/MS integration data of honey bees cuticular hydrocarbon extracts.}
\format{
\subsection{}{

A list with 20 tibble data frames, each with 2 columns:
\describe{
\item{RT}{Retention time measured for each peak within a sample}
\item{Area}{Area result of the integration of the peaks within a sample}
}
}
}
\source{
Data obtained by the author of the package, for educational purposes.
}
\usage{
samples_data_list
}
\description{
Each data frame corresponds tot he integration results for the CHC extract of
a honey bee worker.
}
\keyword{datasets}
