% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_group_table2.R
\name{shape_group_table2}
\alias{shape_group_table2}
\title{Shape the final group table}
\usage{
shape_group_table2(samples_plus_ri)
}
\arguments{
\item{samples_plus_ri}{An aligned data set list containing data frames with
information about compounds and their peak areas and retention times, as
obtained with \link{kovats_retention_index}.}
}
\description{
Shape the final group table
}
