% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_rows.R
\name{shift_rows}
\alias{shift_rows}
\title{Shift rows in a data frame}
\usage{
shift_rows(df, cols_to_shift, rows_to_shift, direction)
}
\arguments{
\item{df}{A data frame to modify.}

\item{cols_to_shift}{A tidy selection of columns to shift.}

\item{rows_to_shift}{A character vector of row names or a numeric vector of row numbers for which the values will be moved.}

\item{direction}{The direction to move the values; either "up" or "down".}
}
\value{
The modified data frame.
}
\description{
Shifts the values in specified columns of a data frame up or down for specified rows.
Rows can be specified using either row names or row numbers.
}
\examples{
# Create a sample data frame with row names
df <- data.frame(col1 = c(1, 2, 3), col2 = c("a", "b", "c"), col3 = c(4, 5, 6))
rownames(df) <- c("row1", "row2", "row3")

# Use the function to shift the values in col1 and col3 one row down for row2 (using row number)
df <- shift_rows(df, c(col1, col3), 2, direction = "down")

}
