% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_comps.R
\name{trace_comps}
\alias{trace_comps}
\title{Filter out compounds of low abundance within samples}
\usage{
trace_comps(unfiltered_group, threshold)
}
\arguments{
\item{unfiltered_group}{An aligned data set list that includes comps.info.}

\item{threshold}{A numeric threshold, indicating the minimum \% of a sample a
peak must represent to remain in the data set.}
}
\description{
The function filters out low-abundance compounds within samples
from an aligned data set list. The user specifies a threshold for
the minimum percentage of a sample that a peak must represent. The
function removes peaks below this threshold and updates the group
area and RT data accordingly. The mean retention time is recalculated
and added to the data.
}
