% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_abundance.R
\name{adjust_abundance}
\alias{adjust_abundance}
\title{Adjust the abundance of the peaks within samples}
\usage{
adjust_abundance(aligned_data, std.info)
}
\arguments{
\item{aligned_data}{Aligned data set, as obtained with \link{recalculate_meanRT}}

\item{std.info}{Data frame with the standards information, as obtained
from \link{shape_hcstd_info}}
}
\description{
Adjust the abundance of the peaks within samples
}
