% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_chromatograms2.R
\name{align_chromatograms2}
\alias{align_chromatograms2}
\title{Aligning peaks based on retention times}
\usage{
align_chromatograms2(
  data2align,
  blanks = NULL,
  linear_shift_criteria,
  partial_alignment_threshold,
  row_merging_threshold
)
}
\arguments{
\item{data2align}{Dataset containing peaks that need to be aligned and matched.}

\item{blanks}{Character vector of names of negative controls.}

\item{linear_shift_criteria}{Numeric value indicating the size range within the linear shift of peaks can
occur during the initial full alignment step.}

\item{partial_alignment_threshold}{Numeric value indicating the threshold for the partial peak alignment step.}

\item{row_merging_threshold}{Numeric valuse indicating the threshold for the merging rows step.}
}
\description{
Custom function for aligning chromatograms.
It is a wrapper around GCalignR::align_chromatograms function, thus it is
recommended to check its documentation (?GCalignR::align_chromatograms).

Its main purpose is to facilitate the usage of the
function with lapply, and within portable scripts.

It sets a seed before applying using GCalignR::align_chromatograms.

The name of the alignment criteria parameters are more similar to their
description in the original paper on the GCalignR package.

rt_col_name is set to "RT" by default, as it is the name given by
import_gcms_data.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
