% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGC-package.R
\docType{package}
\name{analyzeGC-package}
\alias{analyzeGC}
\alias{analyzeGC-package}
\title{analyzeGC: Analysis of GC data}
\description{
A package to facilitate the analysis of Gas-Chromatography (GC) data in R.
}
\keyword{internal}
