% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeGC-package.R
\docType{package}
\name{analyzeGC-package}
\alias{analyzeGC}
\alias{analyzeGC-package}
\title{analyzeGC: Analysis of GC data}
\description{
This package facilitates the analysis of Gas Chromatography (GC) data in R. It was created for educational and research purposes in the field of chemical ecology. The objective is to provide a set of tools for analyzing GC data, particularly GC-MS (Gas Chromatography-Mass Spectrometry) data, in a standardized and reproducible workflow using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dsrodriguezl/analyzeGC}
  \item Report bugs at \url{https://github.com/dsrodriguezl/analyzeGC/issues}
}

}
\keyword{internal}
