% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_df.R
\name{area_df}
\alias{area_df}
\title{Extract aligned area data frames}
\usage{
area_df(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by align_chromatogram2.}
}
\description{
A function to obtain the area aligned data frame from an
aligned data set, as produced by align_chromatogram2.
It automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
