% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_norm.R
\name{area_norm}
\alias{area_norm}
\title{Normalize aligned area data frame}
\usage{
area_norm(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by align_chromatogram2}
}
\description{
Function to normalize the aligned area data frame produced by
align_chromatograms2.

It is a wrapper around GCalignR::norm_peaks.

Its main purpose is to facilitate the usage of the function with lapply.
By default set the rt_col_name to "RT", and conc_col_name to "Area".

Automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
