% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_alignment.R
\name{correct_alignment}
\alias{correct_alignment}
\title{Correct the peaks alignment}
\usage{
correct_alignment(aligned_data, new_peaks = NULL, peak_movements)
}
\arguments{
\item{aligned_data}{Aligned data set as obtained with \link{align_chromatograms2} function.}

\item{new_peaks}{List with instructions for adding empty peaks to the aligned data frames,
if needed.
The default value is NULL, assuming no new peaks should be added.

The names of the list entries correspond to the name of a sample present
in the aligned data frame, to use as reference. The function only add a peak
to the data frame, if finds the reference sample within it.
This facilitates the usage of the function on a list of aligned data frames
with lapply, as the function will add the empty peaks to the correct aligned
data frame within the list.

The entries of the list should correspond to data.frames/tibbles with two
columns ("position.reference" and "direction"). Position.reference
indicates where the empty peak should be added to the aligned data frame.
The position reference indicate a peak (e.g. P10) as a location reference
within the aligned data frame. Direction ("before" or "after") indicates
whether the empty peak should be created as the row before or after the
reference peak.}

\item{peak_movements}{List with instructions to displace the peak area/RT values within the sample
columns of the aligned data frame.

The names of the list entries correspond to the name of the samples for which
the peak area/RT values should be displaced. The function only performs the
indicated peak displacements, if it finds the sample within the given data
set.
This facilitates the usage of the function on a list of aligned data sets
with \link{lapply}, as the function will perform the correct peaks displacements
to the correct data set within the list.

The entries of the list should correspond to data.frames/tibbles with two
columns ("peaks_origin" and "peaks_target").
peaks_origin indicates the peaks (e.g. P10, P12) that holds the values to be
displaced within the aligned data frame. peaks_target indicates the peak to
which which the value should be displaced along the column of the indicated
sample.}
}
\description{
Function to correct alignment of peaks across an aligned data
set as indicated by instructions encoded within data frames' list(s).
It iterates through the data frame samples, following the given instructions,
to add new peaks and displace values from one peak to another
}
\examples{

# Create a list to guide the addition of empty peaks
## Sample 335 is an In-hive worker
## Sample 339 is an Out-hive worker

library(tibble)
empty_peaks <- list("335" = tribble(~position.reference, ~direction,
                                     "P100", "after")
                    , "339" = tribble(~position.reference, ~direction,
                                     "P100", "before"))

# Create a list to guide the displacement of peak values
## Sample 350 is an In-hive worker
## Sample 328 is an Out-hive worker

peaks_movements_list <- list("350" = data.frame(peaks_origin =
                                                      c(paste0("P"
                                                               , c(106, 107
                                                               , 124)))
                                             , peaks_target =
                                             c(paste0("P"
                                                      , c(105, 106
                                                          , 123))))
                        , "328" = data.frame(peaks_origin =
                                                        c(paste0("P"
                                                                 , c(26, 35
                                                                     , 85
                                                                     , 124
                                                                     , 128)))
                                             , peaks_target =
                                               c(paste0("P"
                                                        , c(25, 34
                                                            , 84
                                                            , 123
                                                            , 127)))))
# Correct the alignment of a single aligned area/RT data set
IW <- aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`
IW <- correct_alignment(aligned_data = IW
                             , peak_movements = peaks_movements_list)

# Correct the alignment of several aligned area/RT data frames within a list
corrected_samples_list_area <- lapply(aligned_samples_data_list
                                      , correct_alignment
                                      , peak_movements = peaks_movements_list)


}
