% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_alignment.R
\name{correct_alignment}
\alias{correct_alignment}
\title{Correct the peaks alignment}
\usage{
correct_alignment(aligned_data, movements_list)
}
\arguments{
\item{aligned_data}{Aligned data set as obtained with \link{align_chromatograms2} or
\link{add_empty_peaks} functions.}

\item{movements_list}{List with instructions to displace the peak area/RT values within the sample
columns of the aligned data frame.

The names of the list entries correspond to the name of the samples for which
the peak area/RT values should be displaced. The function only performs the
indicated peak displacements, if it finds the sample within the given data
set.
This facilitates the usage of the function on a list of aligned data sets
with \link{lapply}, as the function will perform the correct peaks displacements
to the correct data set within the list.

The entries of the list should correspond to data.frames/tibbles with two
columns ("peaks_list" and "movement_dirs").
peaks_list indicates the peaks (e.g. P10, P12) that holds the values to be
displaced within the aligned data frame. movement_dirs indicates the
direction ("down" or "up") in which the peak value should be displaced along
the column of the indicated sample.}
}
\description{
Function to correct alignment of peaks across an aligned data
set as indicated by a set of instructions encoded within list of data frames.
It iterates through the data frame samples, following the given instructions,
and calls \link{move_one_peak} to perform each peak displacement.
}
\examples{

# Create a list to guide the displacement of peak values
## Sample 350 is an In-hive worker
## Sample 328 is an Out-hive worker

peaks_movements <- list("350" = data.frame(peaks_list =
                                                      c(paste0("P"
                                                               , c(106, 107
                                                               , 124)))
                                             , movement_dirs = c('up', 'up'
                                                                 , 'up'))
                        , "328" = data.frame(peaks_list =
                                                        c(paste0("P"
                                                                 , c(26, 35
                                                                     , 85
                                                                     , 124
                                                                     , 128)))
                                             , movement_dirs = c('up', 'up'
                                                                 , 'up'
                                                                 , 'up'
                                                                 , 'up')))
# Correct the alignment of a single aligned area/RT data set
IW <- aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`
IW <- correct_alignment(aligned_data = IW
                             , movements_list = peaks_movements)

# Correct the alignment of several aligned area/RT data frames within a list
corrected_samples_list_area <- lapply(aligned_samples_data_list
                                      , correct_alignment
                                      , movements_list = peaks_movements)


}
