% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_alignment.R
\name{correct_alignment}
\alias{correct_alignment}
\title{Correct the peaks alignment}
\usage{
correct_alignment(aligned_data, movements_list)
}
\arguments{
\item{aligned_data}{Aligned data frame as obtained with the area_df or RT_df functions.}

\item{movements_list}{List with instructions to displace the peak area/RT values within the sample
columns of the aligned data frame.

The names of the list entries correspond to the name of the samples for which
the peak area/RT values should be displaced. The function only performs the
indicated peak displacements, if it finds the sample within the given data
frame.
This facilitates the usage of the function on a list of aligned data frames
with lapply, as the function will perform the correct peaks displacements to
the correct data frame within the list.

The entries of the list should correspond to data.frames/tibbles with two
columns ("peaks_list" and "movement_dirs").
Peaks_list indicates the peaks (e.g. P10, P12) that holds the values to be
displaced within the aligned data frame. Movement_dirs indicates the
direction ("down" or "up") in which the peak value should be displaced along
the column of the indicated sample.}
}
\description{
Function to correct alignment of peaks across a data frame, as
indicated by a set of instructions encoded within list of data frames.
It iterates through the data frame samples, following the given instructions,
 and calls move_one_peak to perform each peak displacement.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
