% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_heatmap.R
\name{diagnostic_heatmap}
\alias{diagnostic_heatmap}
\title{Diagnostic}
\usage{
diagnostic_heatmap(df_area_norm, title)
}
\arguments{
\item{df_area_norm}{Normalized area data frame, as produced by area_norm}

\item{title}{Character string with the text for the title of the plot}
}
\description{
Diagnostic
}
