% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_compounds.R
\name{drop_na_compounds}
\alias{drop_na_compounds}
\title{Remove the peaks that do not contain compounds of interest}
\usage{
drop_na_compounds(filtered_data)
}
\arguments{
\item{filtered_data}{An aligned data set list, as obtained with \link{trace_comps}.}
}
\description{
The function takes a list of data frames and removes peaks (rows) that do not
contain compounds of interest. Peaks are considered to not contain compounds
of interest if corresponding entry in the "Compound" column is NA.
}
\examples{
# Removing the NA compounds from a single data set
IW_filtered <- filtered_samples_list$`Winter_In-hive workers_A. m. mellifera`
drop_na_compounds(IW_filtered)

# Removing NA compounds from several data sets within a list
filtered_samples_list2 <- filtered_samples_list |>
  lapply(drop_na_compounds)

}
