% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chc_info.R
\name{get_chc_info}
\alias{get_chc_info}
\title{get the CHC information data frame}
\usage{
get_chc_info(comps.id)
}
\arguments{
\item{comps.id}{data frame with the compound information for the hydrocarbons found in each
sample.

The data frame should have a \strong{"Compound"} and a \strong{"mean_RT"} columns, plus a
column for every sample.

The Compound column should have the name of the hydrocarbons with the
following format:
C\emph{numeric chain length} _ \emph{class abbreviation} _
\emph{numeric position of unsaturations / methyl-branches}.

As an example the entry for the compound 15-Me C20 would be C20_Me_15-.

The function automatically extracts the information of the compound from the
entry in the \strong{"Compound"} column, and replaces it by the correct name.

If a peak contains more than one compound with unsaturations /
methyl-branches, they must be separated by a ";" (e.g. C20_Me_13-;15-
for a peak containing both 13-Me C20 and 15-Me C20).
\subsection{Class abbreviations are accepted by the function:}{

n-alkanes: ane
alkene: ene
alkadiene: diene
monomethyl-branched alkanes: Me
Dimethyl-branched alkanes: Dime
Trimethyl-branched alkanes: Trime

The rest of the data frame, is just as it can be obtained via the \code{\link[=RT_df]{RT_df()}}
function, or after alignment correction.
}}
}
\description{
Function to generate the data frame holding the compound
information for the hydrocarbons present in each sample, within a table
}
