% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mh_integration.R
\name{import_mh_integration}
\alias{import_mh_integration}
\title{Import GC integration results from CSV  files}
\usage{
import_mh_integration(path, patterns_2_delete = " ", zip_export = F)
}
\arguments{
\item{path}{Path of the CSV file with the exported integration results of the
samples.}

\item{patterns_2_delete}{a character type vector, listing the text  patterns
to be removed from the CSV files' names, to extract the names of the samples
from them. This step is performed via str_remove_all.}

\item{zip_export}{Logical value indicating whether to export a ZIP file
with all the separate CSV files of each sample (default: FALSE).}
}
\value{
A list of tibble data frames with the integration results
(RT: retention time of each peak, and Area: area under the curve of each
peak).
Each tibble on the list corresponds to the CSV file of an individual
sample.
}
\description{
A function to import the integration results data frames,
as obtained from MassHunter.
}
