% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_list.R
\name{mg_list}
\alias{mg_list}
\title{group samples within a list}
\usage{
mg_list(sample.info, samples.data.list)
}
\arguments{
\item{samples.data.list}{}
}
\description{
A function to Group samples gcms integration data frames based
on unique group labels, within the list containing the data frames.
}
