% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mg_list.R
\name{mg_list}
\alias{mg_list}
\title{group samples within a list}
\usage{
mg_list(sample.info, group.label, samples.data.list)
}
\arguments{
\item{sample.info}{A data frame indicating the groups to which the samples belong}

\item{group.label}{A character indicating the name of the column of sample.info that contains
the group names by which the samples should be grouped}

\item{samples.data.list}{List of data frames with integration results per sample, as produced by
import_gcms_data}
}
\description{
A function to Group samples gcms integration data frames based
on unique group labels, within the list containing the data frames.
}
\author{
Daniel S. Rodríguez-Leon <72925497+dsrodriguezl@users.noreply.github.com>
}
