% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_meanRT.R
\name{recalculate_meanRT}
\alias{recalculate_meanRT}
\title{recalculate the mean RT of data set after alignment correction}
\usage{
recalculate_meanRT(aligned_data)
}
\arguments{
\item{aligned_data}{Aligned data set as obtained with \link{correct_alignment}}
}
\description{
Function to correct the mean RT values of an aligned data frame. It is
called within the correct_alignment function to correct the mean RT values
of the peaks after moving the peak values within the samples, as part of the
alignment correction process.
}
