% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_meanRT.R
\name{recalculate_meanRT}
\alias{recalculate_meanRT}
\title{recalculate the mean RT of a list of data frames after alignment
correction}
\usage{
recalculate_meanRT(RT.list, area.list, output.list)
}
\arguments{
\item{RT.list}{List containing the RT aligned data frames after alignment correction with
correct_alignment.}

\item{area.list}{List containing the area aligned data frames after alignment correction with
correct_alignment.}

\item{output.list}{A character string indicating which of the lists ("Area" or "RT") by the
function, after adding to its data frames the newly calculated mean RTs.}
}
\description{
recalculate the mean RT of a list of data frames after alignment
correction
}
\examples{

corrected_samples_list_area2 <-
  recalculate_meanRT(RT.list = corrected_samples_list_RT
                     , area.list = corrected_samples_list_area
                     , output.list = "Area")

corrected_samples_list_RT2 <-
  recalculate_meanRT(RT.list = corrected_samples_list_RT
                     , area.list = corrected_samples_list_area
                     , output.list = "RT")


}
