% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculate_meanRT.R
\name{recalculate_meanRT}
\alias{recalculate_meanRT}
\title{recalculate the mean RT of data set after alignment correction}
\usage{
recalculate_meanRT(aligned_data)
}
\arguments{
\item{aligned_data}{Aligned data set as obtained with \link{correct_alignment}}
}
\description{
recalculate the mean RT of data set after alignment correction
}
\examples{

# Recalculate the mean RT for a single data set
corrected_IW <- corrected_samples_list$`Winter_In-hive workers_A. m. mellifera`

corrected_IW <-recalculate_meanRT(corrected_IW)

# Recalculate the mean RT for several data sets within a list
corrected_samples_list2 <- lapply(corrected_samples_list, recalculate_meanRT)

}
