% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_df.R
\name{RT_df}
\alias{RT_df}
\title{Extract aligned retention time data frames}
\usage{
RT_df(aligned_data)
}
\arguments{
\item{aligned_data}{An aligned data set as produced by align_chromatogram2}
}
\description{
A function to obtain the retention time (RT) aligned data frame
from an aligned data set, as produced by align_chromatogram2.
It automatically labels each peak as P1 to Pn, where n  =  number of peaks
in the data set.
}
\examples{

# Extract area data frame from a single aligned data set
Winter_IW <-
  aligned_samples_data_list$`Winter_In-hive workers_A. m. mellifera`

RT_winter_IW <- RT_df(Winter_IW)

# Extract area data frame from several aligned data sets within a list
samples_list_RT <- aligned_samples_data_list |>
  lapply(RT_df)

}
