% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_hcstd_info.R
\name{shape_hcstd_info}
\alias{shape_hcstd_info}
\title{Shape the standards information data frame}
\usage{
shape_hcstd_info(comps_id.STD, std_area, short_std_pattern, long_std_pattern)
}
\arguments{
\item{comps_id.STD}{data frame with the compound information for the n-alkane standards.

The data frame should have a "Compound" and a "mean_RT" columns, plus a
column for every standard run.

The Compound column should have the name of the n-alkane standards with the
following format: C"numeric chain length"_ane_NA.
The rest of the data frame, is just as it can be obtained via the RT_df
function, or after alignment correction.}

\item{std_area}{Aligned area data frame for the standards, as it is obtained with area_df,
or after alignment correction.}

\item{short_std_pattern}{Character string pattern to identify the short chain-length standards (until
C20). The function relies on the dplyr::starts_with selection helper to do
so.}

\item{long_std_pattern}{Character string pattern to identify the long chain-length standards (from
C21). The function relies on the dplyr::starts_with selection helper to do
so.}
}
\description{
Function to shape the data frame holding the compound
information for n-alkane standards
}
