% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_comps.R
\name{trace_comps}
\alias{trace_comps}
\title{Filter out compounds of low abundance within samples}
\usage{
trace_comps(unfiltered_group, threshold)
}
\arguments{
\item{unfiltered_group}{An aligned data set list that includes comps.info.}

\item{threshold}{A numeric threshold, indicating the minimum \% of a sample a
peak must represent to remain in the data set.}
}
\description{
Filter out compounds of low abundance within samples
}
